+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 
 ### Additions
 * Proper `has_many` and `has_one` associations
-* `AIXM::Association:Array#find` and `AIXM::Association:Array#duplicate` on
-  `has_many` associations.
+* `AIXM::Association:Array#find_by|find|duplicate` on `has_many` associations
 * `AIXM.config.mid` now defines whether `mid` attributes are inserted or not
   provided the selected schema is OFMX
 * `AIXM::Memoize` module